Update uncertainties.ipynb
"Fixed? broken link to Populations user guide. (29)
(unable to see in notebook form, only as the raw code representation; but, this code snip was taken from the 'Curating' page's code that was working so I expect that it is also working here as a markdown link!)."


Fixed spelling: uncertaintoies -> uncertainties. (102)


Fixed spelling: associed -> associated. (9)